Refactor settings components as functional in separate file
